mudei int->float

+++ study_doc/alura/soma_user.py
@@ -1,6 +1,6 @@
 #some de dois números digitados pelo usuário
-primeiro_numero_user=int(input('Digite um número: '))
-segundo_numero_user=int(input('Digite um número: '))
+primeiro_numero_user=float(input('Digite um número: '))
+segundo_numero_user=float(input('Digite um número: '))
 
 soma=sum((primeiro_numero_user,segundo_numero_user))
